Introduce module to generate group generators for schottky groups
This module generates the the bounding geodesics for a schottky group
with any number of generators. Then the bounding geodesics are used to
compute the group generators.

Attention: The circles are placed equidistantly and opposite circles
are identified. This leads to highly symmetrical groups. If this is
not wanted, the module has to be updated.